checkpoint #4 -> next part refactor (Forum part)

+++ HarmonyHomeNet-Logic/src/main/java/bwp/pwr/daniel/rysz/harmonyhomenetlogic/mainLogicEntitys/forum/service/ForumServiceImp.java
@@ -1,12 +1,14 @@
 package bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.forum.service;
 
+import bwp.pwr.daniel.rysz.harmonyhomenetlogic.exeptions.customErrors.ForumNotFoundException;
 import bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.forum.entity.Forum;
 import bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.forum.repository.ForumRepository;
+import bwp.pwr.daniel.rysz.harmonyhomenetlogic.utils.response.forumStaff.ForumResponse;
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -21,22 +23,42 @@ public class ForumServiceImp implements ForumService {
     }
 
     @Override
-    public void save(Forum forum) {
+    public ForumResponse save(@NonNull Forum forum) {
         forumRepository.save(forum);
+        return mapForumToForumResponse(forum);
     }
 
     @Override
-    public Optional<Forum> findById(UUID id) {
-        return forumRepository.findById(id);
+    public Forum findById(UUID id) throws ForumNotFoundException {
+        return forumRepository.findById(id)
+                .orElseThrow(() -> new ForumNotFoundException("wrong forum id"));
     }
 
     @Override
-    public Optional<Forum> findByForumName(String forumName) {
-        return forumRepository.findByForumName(forumName);
+    public Forum findByForumName(String forumName) throws ForumNotFoundException {
+        return forumRepository.findByForumName(forumName)
+                .orElseThrow(() -> new ForumNotFoundException("wrong forum name"));
     }
 
     @Override
-    public void deleteById(UUID id) {
-        forumRepository.deleteById(id);
+    public void deleteById(UUID id) throws ForumNotFoundException {
+        if (forumRepository.existsById(id)) forumRepository.deleteById(id);
+        else throw new ForumNotFoundException("wrong forum id");
+    }
+
+    @Override
+    public ForumResponse mapForumToForumResponse(@NonNull Forum forum) {
+        return ForumResponse.builder()
+                .id(forum.getId())
+                .forumName(forum.getForumName())
+                .forumDescription(forum.getForumDescription())
+                .build();
+    }
+
+    @Override
+    public List<ForumResponse> mapForumListToForumResponseList(@NonNull List<Forum> forums) {
+        return forums.stream()
+                .map(this::mapForumToForumResponse)
+                .toList();
     }
 }
